action_server: add action server to resemble move_base actions
This actionserver is supposed to be used on top of the move_base in
order to allow checking of global paths for possible collisions.

In case a path calculated by move_base's used global planner would
collide with another robot, the velocity_share_actionserver sends a new
goal in front of that collision.
When the path gets free again the actionserver advances to the original
goal.